ToughResources
-Движение камеры на среднюю кнопку мыши
-Скрипт ToughResources, который лежит на объектах, которые добываются рабочими(Деревья и рудники и тд)
-Обновлён скрипт Worker, рабочие могут тратить действия на то, чтобы добывать ресурсы
-Добавлен префаб дерева
-Fighter теперь может ходить только по пустым клеткам и подбираемым ресурсам(CollectableItem)
-Все short, byte и тд заменены на int

+++ Assets/Scripts/UnitDescription.cs
@@ -8,11 +8,11 @@ public class UnitDescription : ObjectOnGrid
     public List<float> ArmorEfficiencyTable; // Таблица, где номер элемента - количество брони, а значение элемента - соответствующая ей %-ая защита от урона
     public bool IsSelected;
     public bool IsHighlighted;
-    public byte ActionsPerTurn;
+    public int ActionsPerTurn;
     public int Armor;
     public int TeamAffiliation;
-    public short MovementSpeed;
-    public short AttackRange;
+    public int MovementSpeed;
+    public int AttackRange;
     public float Health;
     public float AttackDamage;
     public float FoodConsumption;
